v7.6.1
XItemStack
If the item name is empty, it'll assume that the user wants the item to have no name.
Fixed Shulker Box support in 1.14 and below.
Fixed item attributes. They're only supported in 1.13+

ParticleDisplay
Added toString() method.
Added support for displaying the particle to specific players only.
Added an option to eedit an existing ParticleDisplay properties only from the config values that are different.