Added serialization support for Maps
Fixed NPE when passing AIR (or some other item for which getItemMeta() method returns null) as parameter of serialize method.

Fixed NPE when serializing items that had AttributeModifiers but didn't specify a slot.

Fixed NPE when deserializing items that had AttributeModifier, that was happening because instead of querying the specific config entry of that Attribute, the code was querying the map of attributes, which of course does not have any attribute of specific properties on its root.