v8.6.0
Unit tests are now much more complete.

XTag
A class similar to Spigots XTag, but with more features and directly works with XSeries.
Note that this is a huge class. It's recommended that you exclude this from your project if you don't want to use it.

MaterialUtils
Removed in favor of XTag.

ReflectionUtils
Fixed an issue with 1.17
Attempted to fix a few issues with Bukkit-Forge servers.

XMaterial
Fixed an issue with Regex cache.

XSound
Added a method to stop a particular sound from XSound.Record

NMSExtras
Added support for 1.18 and fixed a small issue for 1.17

XPotion
If you're using this utility, you should consider recoding all those parts.
Most methods are renamed to a shorter and concise version.
Added a wrapper for PotionEffect for chances.
matchXPotion(PotionEffectType) is now extremely more efficient.